% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berraerkosar-package.R
\docType{package}
\name{berraerkosar-package}
\alias{berraerkosar}
\alias{berraerkosar-package}
\title{berraerkosar: Dev-Functions}
\description{
This is an R package that contains two functions. The first one calculates a body mass index. And the second one takes as parameter a single file with the format ‘xlsx’ and returns an object of format list containing the different tables of each tab of the file.
}
\author{
\strong{Maintainer}: Berra Erkosar \email{berra.erkosar@finddx.org} (\href{https://orcid.org/0000-0002-1565-9316}{ORCID})

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
