% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{BMI}
\format{
A data frame with masses and heights
}
\source{
embedded data
}
\usage{
bmi(mass, height)
}
\arguments{
\item{mass}{Enter the weight in kg}

\item{height}{Enter the height in meters}
}
\value{
Returns the body mass index (BMI)
}
\description{
BMI
}
\examples{
library(dplyr)

bmi(mass = 80, height = 1.80) # Should give 24.7

#To insert BMI column on a data frame
data("data_weight")
data_weight \%>\%
  mutate(bmi = bmi(mass = mass, height = height))
}
