% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_weight.R
\docType{data}
\name{data_weight}
\alias{data_weight}
\title{Weights and Heights
A dataset containing the height and weight data}
\format{
A data frame with 87 rows and 3 columns: #' \describe{
\item{name}{name, a name}
\item{height}{height in meters}
\item{mass}{mass in kg}
#' ...
#' }
@source \url{www.thinkr.com}
}
\usage{
data_weight
}
\description{
Weights and Heights
A dataset containing the height and weight data
}
\keyword{datasets}
