% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_import_excel.R
\name{multi_import_excel}
\alias{multi_import_excel}
\title{multi_import_excel}
\usage{
multi_import_excel(filename)
}
\arguments{
\item{filename}{The excel file}
}
\value{
list of sheets
}
\description{
multi_import_excel
}
\examples{
datafile <- system.file("toydatasets.xlsx", package = "berraerkosar")
multi_import_excel(datafile)
}
